testing

+++ application/views/User/View_sheet_Invoice.php
@@ -491,9 +491,9 @@
 
     // No of  Pieces
     $("#no_holes").on('change keyup paste', function() {
-        var holes = parseInt($('#no_holes').val());
-        var amt = parseInt($('#charge_amt').val());
-        var total =  parseInt(holes * amt);
+        var holes = parseFloat($('#no_holes').val());
+        var amt = parseFloat($('#charge_amt').val());
+        var total =  parseFloat(holes * amt);
         document.getElementById('tot_charge_amt').value = total;
 
         var totals =document.getElementsByName("tot_charge_amt[]");
@@ -557,9 +557,9 @@
 
     // Charge Amount
     $("#charge_amt").on('change keyup paste', function() {
-        var amt = parseInt($('#charge_amt').val());
-        var holes = parseInt($('#no_holes').val());
-        var total =  parseInt(holes * amt);
+        var amt = parseFloat($('#charge_amt').val());
+        var holes = parseFloat($('#no_holes').val());
+        var total =  parseFloat(holes * amt);
         document.getElementById('tot_charge_amt').value = total;
 
         var totals =document.getElementsByName("tot_charge_amt[]");
